Cleanup and fixes
Loads of comments
general code cleanup
Fixed unloading of left players
Added arguments to dedicated server for starting
  gameserver and AI separately
Fixed missing text update of loading screen waitbar
Enhanced initial start player selection

+++ src/room/DRoomAI.py
@@ -1,3 +1,13 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+__author__ = "Fireclaw the Fox"
+__license__ = """
+Simplified BSD (BSD 2-Clause) License.
+See License.txt or http://opensource.org/licenses/BSD-2-Clause for more info
+"""
+
+import random
+
 from direct.distributed.DistributedObjectAI import DistributedObjectAI
 from dice.SixSidedDice import SixSidedDice
 from globalData import RoomGlobals
@@ -40,8 +50,13 @@ class DRoomAI(DistributedObjectAI):
     def delete(self):
         self.air.roomZoneAllocator.free(self.roomZone)
         self.roomZone = None
+
+        # cleanup players
+        for player in self.playerList:
+            self.air.sendDeleteMsg(player)
         self.playerList = []
 
+        # cleanup board
         self.air.sendDeleteMsg(self.boardAI.doId)
 
         DistributedObjectAI.delete(self)
@@ -60,17 +75,12 @@ class DRoomAI(DistributedObjectAI):
 
         # add the player to the list
         self.playerList.append(player)
-        if self.activePlayerId == -1:
-            # the first joint player will be the starting player
-            #TODO: This may need to be changed for other game types like race
-            self.activePlayerId = self.playerList[0].avId
 
         # find a start position for the new player
         startField = self.boardAI.getStartField(self.getPlayerCount())
         if startField == None:
             print("Room {} couldn't find start position".format(self.name))
             return False
-        #print("Set position to {}".format(startPos))
         player.currentField = startField
         player.startField = startField
         player.piece.setPos(startField.nodepath.getPos())
@@ -81,10 +91,13 @@ class DRoomAI(DistributedObjectAI):
         return True
 
     def endTurn(self):
+        """End the current players turn. Can only be called by the player whos
+        turn it actually is at the moment."""
         playerId = self.air.getAvatarIdFromSender()
         if playerId != self.activePlayerId:
             return
 
+        # reset the roll so the next player can roll the dice again
         self.currentRoll = -1
 
         nextPlayer = None
@@ -109,16 +122,23 @@ class DRoomAI(DistributedObjectAI):
         self.sendUpdateToAvatarId(nextPlayer.avId, "startTurn", [])
 
     def isFull(self):
+        """Returns wether there are already the maximum amount of players in
+        this room or not"""
         return self.getPlayerCount() == self.maxPlayers
 
     def d_startRoom(self):
-        print("TRY START ROOM!")
+        """Tell the players that the room is starting."""
         # we can only start if all players are ready
         if len(self.readyPlayers) != self.maxPlayers: return
 
-        playerId = self.air.getAvatarIdFromSender()
-        print("PLAYER ID START ROOM:", playerId)
-        print("START ROOM!")
+        # determine who will be the starting player
+        if self.activePlayerId == -1:
+            if self.gameType == RoomGlobals.GAMETYPE_RACE:
+                # on race mode, choose a random player
+                self.activePlayerId = random.choice(self.playerList).avId
+            else:
+                # the first joint player will be the starting player
+                self.activePlayerId = self.playerList[0].avId
 
         nextPlayer = None
         for player in self.playerList:
@@ -126,41 +146,42 @@ class DRoomAI(DistributedObjectAI):
             nextPlayer = player
             break
 
-        print("UPDATE ALL PLAYERS")
         # send an update to the channel so everyone knows who's next
         self.sendUpdate("nextPlayer", [nextPlayer.getName()])
 
-        print("SEND UPDATE TO:", self.activePlayerId)
         # send the start turn to the next active player so he can enable the gui
         # and other relevant elements
         self.sendUpdateToAvatarId(nextPlayer.avId, "startTurn", [])
 
-        print("START ROOM FROM AI")
         self.sendUpdate("startRoom", [])
 
     def playerReady(self, playerId):
-        print("GOT PLAYER READY!", playerId)
         if playerId not in self.readyPlayers:
             self.readyPlayers.append(playerId)
 
         self.d_startRoom()
 
     def removePlayer(self, playerId):
+        """Remove the player with the given ID from the rooms player list"""
+        self.air.sendDeleteMsg(playerId)
+
         for p in self.playerList[:]:
             if p.doId == playerId:
                 if self.activePlayerId == p.avId:
                     self.endTurn()
                 self.playerList.remove(p)
-        self.air.sendDeleteMsg(playerId)
         return len(self.playerList) == 0
 
     def getPlayerNames(self):
+        """Return a list of all player names in this room"""
         names = []
         for p in self.playerList:
             names.append(p.name)
         return names
 
     def rollDice(self):
+        """Roll the dice and store the value. This can only be called by the
+        active player"""
         playerId = self.air.getAvatarIdFromSender()
         if playerId != self.activePlayerId:
             self.sendUpdateToAvatarId(playerId, "rolledDiceFailed", [])
@@ -173,6 +194,9 @@ class DRoomAI(DistributedObjectAI):
             self.sendUpdateToAvatarId(playerId, "rolledDiceFailed", [])
 
     def requestMoveToField(self, fieldName):
+        """Request to move the player to the given field. This can onyl be
+        called by the active player and will be refused if the players roll is
+        not sufficient to move to the given field"""
         playerId = self.air.getAvatarIdFromSender()
         if playerId != self.activePlayerId:
             return
@@ -208,6 +232,7 @@ class DRoomAI(DistributedObjectAI):
             self.sendUpdateToAvatarId(playerId, "updateRolledDice", [self.currentRoll])
 
     def setMultiplePlayersOnField(self, field):
+        """Move the players a bit to not position them inside each other"""
         playersOnField = []
 
         for player in self.playerList:
@@ -224,18 +249,23 @@ class DRoomAI(DistributedObjectAI):
             playersOnField[0].moveTo(field, 0, 0)
 
     def canInitiateFight(self, field):
-        print("CAN INITIATE ROOM AI")
+        """A player can start a fight on the given field. Tell that to the
+        player so he can decide what to do."""
         playerId = self.air.getAvatarIdFromSender()
         self.canBattle = True
         self.currentFightField = field
         self.sendUpdateToAvatarId(playerId, "canInitiateFight", [])
 
     def initiateDirectFight(self, field):
+        """Skip the questioning of players if they want to start a fight and
+        directly jump into it."""
         self.canBattle = True
         self.currentFightField = field
         self.initiateFight()
 
     def initiateFight(self):
+        """Start a fight on the previously set fight field if a battle can be
+        initiated."""
         if not self.canBattle: return
         self.canBattle = False
         field = self.currentFightField
@@ -246,8 +276,6 @@ class DRoomAI(DistributedObjectAI):
         for player in self.playerList:
             if player not in playersOnField:
                 spectatorPlayers.append(player)
-
-        print("START FIGHT WITH:", playersOnField)
 
 
         self.battleAI = DBattleAI(self.air, field, playersOnField, spectatorPlayers, self.difficulty)
@@ -262,31 +290,32 @@ class DRoomAI(DistributedObjectAI):
             self.sendUpdateToAvatarId(player.avId, "spectateBattle", [])
 
 
-        print("ACCEPT", self.battleAI.uniqueName("endBattle"))
         self.accept(self.battleAI.uniqueName("endBattle"), self.stopFight)
 
     def stopFight(self, field, playersWon):
-        print("BATTLE OVER, DELETE OBJECT")
+        """Called when the battle is over, given the fight the battle was held
+        on and a boolean value wheter the players or the enemies have won."""
         if playersWon:
             self.boardAI.wonFight(field)
         else:
-            print("PLAYERS LOST SO RESET THEIR POSITION")
-            print(self.battleAI.playersAttending)
             for player in self.battleAI.playersAttending:
-                print("RESETTING PLAYER: ", player)
                 player.resetToStartField()
 
         # cleanup the battlefield
-        self.air.deleteObject(self.battleAI.doId)
-        del self.battleAI
+        if hasattr(self, "battleAI"):
+            self.air.deleteObject(self.battleAI.doId)
+            del self.battleAI
 
     def levelUpAllPlayers(self):
+        """Raise the level of all players by 1 if they can level up. Currently
+        capped at level 2"""
         for player in self.playerList:
             if player.level == 1:
                 player.level += 1
                 player.updateInventory()
 
     def gameOver(self, field):
+        """The game is over, tell weveryone who won this round"""
         if self.gameType == RoomGlobals.GAMETYPE_RACE:
             winningPlayerName = ""
             for player in self.playerList:
